refactor: create new classes TemplateDrawing and SaveImageOps
