+++ src/processors/internal/CropOnPatchesCommon.py
@@ -12,7 +12,7 @@ from src.processors.constants import (
 )
 from src.processors.internal.WarpOnPointsCommon import WarpOnPointsCommon
 from src.utils.constants import CLR_DARK_GREEN
-from src.utils.image import ImageUtils
+from src.utils.drawing import DrawingUtils
 from src.utils.logger import logger
 from src.utils.math import MathUtils
 from src.utils.parsing import OVERRIDE_MERGER
@@ -211,16 +211,16 @@ class CropOnPatchesCommon(WarpOnPointsCommon):
         if len(area_control_points) > 1:
             if len(area_control_points) == 2:
                 # Draw line if it's just two points
-                ImageUtils.draw_contour(self.debug_image, area_control_points)
+                DrawingUtils.draw_contour(self.debug_image, area_control_points)
             else:
                 # Draw convex hull of the found control points
-                ImageUtils.draw_contour(
+                DrawingUtils.draw_contour(
                     self.debug_image,
                     cv2.convexHull(np.intp(area_control_points)),
                 )
 
         # Helper for alignment
-        ImageUtils.draw_arrows(
+        DrawingUtils.draw_arrows(
             self.debug_image,
             area_control_points,
             area_destination_points,
@@ -228,7 +228,7 @@ class CropOnPatchesCommon(WarpOnPointsCommon):
         )
         for control_point in area_control_points:
             # Show current detections too
-            ImageUtils.draw_box(
+            DrawingUtils.draw_box(
                 self.debug_image,
                 control_point,
                 # TODO: change this based on image shape
@@ -314,7 +314,7 @@ class CropOnPatchesCommon(WarpOnPointsCommon):
 
         config = self.tuning_config
         if config.outputs.show_image_level >= 1:
-            ImageUtils.draw_box_diagonal(
+            DrawingUtils.draw_box_diagonal(
                 self.debug_image,
                 area_start,
                 area_end,
